Mover método de una clase a otra
